Safari-611.1.10/LayoutTests/fast/forms/file/file-input-capture-expected.txt

Tests the capture attribute of <input type='file'>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'capture' in input is true
PASS input.capture is false
PASS input.hasAttribute('capture') is false
PASS input.capture is false
PASS input.hasAttribute('capture') is false
PASS input.capture is true
PASS input.hasAttribute('capture') is true
PASS input.capture is false
PASS input.hasAttribute('capture') is false
PASS input.capture is true
PASS input.hasAttribute('capture') is true
PASS input.capture is false
PASS input.hasAttribute('capture') is false
PASS input.capture is true
PASS input.hasAttribute('capture') is true
PASS successfullyParsed is true

TEST COMPLETE
